refactor: move staking and xyk implementations into separate files
